+++ src/interface/rodeo_resolver.rs
@@ -1,0 +1,28 @@
+//! Implementations of [`Resolver`] for [`RodeoResolver`]
+
+use crate::{Key, Resolver, RodeoResolver};
+
+impl<K> Resolver<K> for RodeoResolver<K>
+where
+    K: Key,
+{
+    #[cfg_attr(feature = "inline-more", inline)]
+    fn resolve<'a>(&'a self, key: &K) -> &'a str {
+        self.resolve(key)
+    }
+
+    #[cfg_attr(feature = "inline-more", inline)]
+    fn try_resolve<'a>(&'a self, key: &K) -> Option<&'a str> {
+        self.try_resolve(key)
+    }
+
+    #[cfg_attr(feature = "inline-more", inline)]
+    unsafe fn resolve_unchecked<'a>(&'a self, key: &K) -> &'a str {
+        self.resolve_unchecked(key)
+    }
+
+    #[cfg_attr(feature = "inline-more", inline)]
+    fn contains_key(&self, key: &K) -> bool {
+        self.contains_key(key)
+    }
+}